move loop to resources rather than construct
